fix 1.8 does not have swap item event
casue error like
[09:48:37 ERROR]: [UltraCosmetics] Plugin UltraCosmetics v2.1.3 has failed to register events for class be.isach.ultracosmetics.cosmetics.gadgets.Gadget$GadgetListener because org/bukkit/event/player/PlayerSwapHandItemsEvent does not exist.

and all listeners stop working
